Desktop/RFID Borrowing system/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<5.0
qrcode>=7.4
Pillow>=10.0
django-extensions>=3.2.0
pyopenssl>=23.0.0
werkzeug>=2.3.0


